move conditional sampler in statistical tools
